timescaledb-project/requirements.txt
psycopg2-binary
psutil
streamlit
pandas
